+++ pkg/local_object_storage/metabase/version_test.go
@@ -1,6 +1,7 @@
 package meta
 
 import (
+	"bytes"
 	"encoding/binary"
 	"errors"
 	"fmt"
@@ -13,13 +14,16 @@ import (
 
 	"github.com/nspcc-dev/bbolt"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
+	"github.com/nspcc-dev/neofs-sdk-go/checksum"
 	checksumtest "github.com/nspcc-dev/neofs-sdk-go/checksum/test"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
+	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
 	objecttest "github.com/nspcc-dev/neofs-sdk-go/object/test"
 	usertest "github.com/nspcc-dev/neofs-sdk-go/user/test"
+	"github.com/nspcc-dev/tzhash/tz"
 	"github.com/stretchr/testify/require"
 )
 
@@ -628,4 +632,120 @@ func TestMigrate9To10(t *testing.T) {
 
 		return nil
 	}))
+}
+
+//nolint:staticcheck // the whole tests is about checking deprecated values
+func TestMigrate10To11(t *testing.T) {
+	var (
+		db   = newDB(t)
+		cID1 = cidtest.ID()
+		cID2 = cidtest.ID()
+	)
+
+	const numOfTestObjs = 2005 // a little more than single iteration in `updateContainersInterruptable` for two containers
+	objs := make([]object.Object, 0, numOfTestObjs)
+	for i := range numOfTestObjs {
+		o := objecttest.Object()
+		o.SetPayloadHomomorphicHash(checksum.NewTillichZemor([tz.Size]byte{}))
+		if i < numOfTestObjs {
+			o.SetContainerID(cID1)
+		} else {
+			o.SetContainerID(cID2)
+		}
+
+		objs = append(objs, o)
+	}
+
+	err := db.boltDB.Update(func(tx *bbolt.Tx) error {
+		bkt1, err := tx.CreateBucketIfNotExists(metaBucketKey(cID1))
+		require.NoError(t, err)
+		bkt2, err := tx.CreateBucketIfNotExists(metaBucketKey(cID2))
+		require.NoError(t, err)
+
+		for _, o := range objs {
+			err = PutMetadataForObject(tx, o, true)
+			if err != nil {
+				return err
+			}
+		}
+
+		for i, o := range objs {
+			var bkt *bbolt.Bucket
+			if i < numOfTestObjs/2 {
+				bkt = bkt1
+			} else {
+				bkt = bkt2
+			}
+
+			// copied from old `PutMetadataForObject` version with homomorphic hashes
+			{
+				var keyBuf keyBuffer
+				if h, ok := o.PayloadHomomorphicHash(); ok {
+					if err = putPlainAttribute(bkt, &keyBuf, o.GetID(), object.FilterPayloadHomomorphicHash, string(h.Value())); err != nil {
+						return err
+					}
+				}
+			}
+		}
+
+		return nil
+	})
+	require.NoError(t, err)
+
+	countFields := func(db *bbolt.DB) (int, error) {
+		var numOfFields int
+		err := db.View(func(tx *bbolt.Tx) error {
+			for _, cID := range []cid.ID{cID1, cID2} {
+				b := tx.Bucket(metaBucketKey(cID))
+				err = b.ForEach(func(_, _ []byte) error {
+					numOfFields++
+					return nil
+				})
+				if err != nil {
+					return err
+				}
+			}
+			return nil
+		})
+		if err != nil {
+			return 0, err
+		}
+
+		return numOfFields, nil
+	}
+
+	numOfFieldsBefore, err := countFields(db.boltDB)
+	require.NoError(t, err)
+
+	err = updateContainersInterruptable(db, []byte{metadataPrefix}, dropHomomorphicIndexes)
+	require.NoError(t, err)
+
+	numOfFieldsAfter, err := countFields(db.boltDB)
+	require.NoError(t, err)
+
+	require.Equal(t, numOfFieldsBefore-2*numOfTestObjs, numOfFieldsAfter) // two indexes deleted for every object
+
+	err = db.boltDB.View(func(tx *bbolt.Tx) error {
+		for _, cID := range []cid.ID{cID1, cID2} {
+			b := tx.Bucket(metaBucketKey(cID))
+			c := b.Cursor()
+
+			for k, _ := c.First(); k != nil; k, _ = c.Next() {
+				switch k[0] {
+				case metaPrefixAttrIDPlain:
+					if bytes.HasPrefix(k[1:], []byte(object.FilterPayloadHomomorphicHash)) {
+						return fmt.Errorf("found ATTR -> ID key for %s container: %x", cID, k)
+					}
+				case metaPrefixIDAttr:
+					if bytes.HasPrefix(k[1+oid.Size:], []byte(object.FilterPayloadHomomorphicHash)) {
+						return fmt.Errorf("found ID -> ATTR key for %s container: %x", cID, k)
+					}
+				default:
+				}
+			}
+		}
+
+		return nil
+	})
+	require.NoError(t, err)
 }